+++ tests/testthat/test-utilities.R
@@ -2,7 +2,7 @@ library(texPreview)
 
 testthat::context('utility functions of package')
 
- testthat::skip_on_cran()
+ #testthat::skip_on_cran()
 
 testthat::describe('build usepackage call',{
  
@@ -47,6 +47,8 @@ testthat::describe('build usepackage multiple calls',{
 })
 
 testthat::describe('find packages',{
+  
+  skip_on_windows()
   
   if(length(system('tlmgr --version',intern = TRUE))>0){